+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.8.2]
+
+### Changed
+- improved process handling
+- minor settings dialog adjustments
+- updated test-resources
+- Gradle-wrapper update
+- other dependency updates
+
 ## [1.8.1]
 
 ### Fixed